## [2.0.9] - 2017-12-29 ### Changed - Optimization for parser
### Fixed
- Bug name permission for pages
