+++ src/Platform/Http/Composers/MenuComposer.php
@@ -197,7 +197,7 @@ class MenuComposer
                 'route'      => route('dashboard.pages.show', [$page->slug]),
                 'label'      => $page->name,
                 'childs'     => false,
-                'permission' => 'dashboard.posts.type.' . $page->slug,
+                'permission' => 'dashboard.pages.type.' . $page->slug,
                 'sort'       => $key,
                 'groupname'  => $page->groupname,
                 'divider'    => $page->divider,
