+++ src/Platform/Http/Controllers/Posts/PageController.php
@@ -29,7 +29,6 @@ class PageController extends Controller
      * @param Page $page
      *
      * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
-     * @throws \Orchid\Platform\Exceptions\TypeException
      */
     public function show(Page $page = null)
     {
@@ -47,7 +46,6 @@ class PageController extends Controller
      * @param Request $request
      *
      * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
-     * @throws \Orchid\Platform\Exceptions\TypeException
      */
     public function update(Page $page, Request $request)
     {
